// vibe/overlay/extensions/vibe-chandra/src/model/types.ts
export interface Hyperedge {
	id: string;
	sources: string[];
	target: string;
}